refactor: usersController api  getUser and putUser
